Fixed stuff, tweaked Error-404 zoom and added dracorock
Fixed clipping issue on fields hopefully for bf. Fixed freeplay crash when exiting a song selected from there. Moved down and in both opponents and fields. Moved them in specifically to match up more with the olevadon week ref and to allow to cut more from the left to fix the clipping issue. Adjusted Error-404 zoom.